hw_2_workspace/homework_2_2020/src/it/uniroma3/diadia/giocatore/Borsa.java
/**
 * 
 */
package it.uniroma3.diadia.giocatore;

import it.uniroma3.diadia.attrezzi.Attrezzo;

/**
 * Class Borsa - Contiene gli attrezzi del giocatore.
 * 
 * @author docente di POO
 *
 */
public class Borsa {
	
	public final static int DEFAULT_PESO_MAX_BORSA = 10;
	public final static int NUMERO_MASSIMO_ATTREZZI = 10;
	private Attrezzo[] attrezzi;
	private int numeroAttrezzi;
	private int pesoMax;
	
	public Borsa() {
		this(DEFAULT_PESO_MAX_BORSA);
	}
	
	public Borsa(int pesoMax) {
		this.pesoMax = pesoMax;
		this.attrezzi = new Attrezzo[10]; // speriamo che bastino...
		this.numeroAttrezzi = 0;
	}

	public boolean addAttrezzo(Attrezzo attrezzo) {
		if (this.getPeso() + attrezzo.getPeso() > this.getPesoMax())
			return false;
		if (this.numeroAttrezzi==10)
			return false;
		this.attrezzi[this.numeroAttrezzi] = attrezzo;
		this.numeroAttrezzi++;
		return true;
	}

	public int getPesoMax() {
		return pesoMax;
	}

	public Attrezzo getAttrezzo(String nomeAttrezzo) {
		Attrezzo attrezzoCercato = null;
		for(Attrezzo attrezzo: this.attrezzi) { 
			if (attrezzo==null)
				continue;
			if (attrezzo.getNome().equals(nomeAttrezzo))
				attrezzoCercato = attrezzo;
		}
		return attrezzoCercato;
	}

	public int getPeso() {
		int peso = 0;
		for (int i= 0; i<this.numeroAttrezzi; i++)
			peso += this.attrezzi[i].getPeso();
		return peso;
	}

	public boolean isEmpty() {
		return this.numeroAttrezzi == 0;
	}

	public boolean hasAttrezzo(String nomeAttrezzo) {
		return this.getAttrezzo(nomeAttrezzo)!=null;
	}

	public Attrezzo removeAttrezzo(String nomeAttrezzo) {
		Attrezzo attrezzo = null;
		int i=0;
		while(attrezzo==null && i < NUMERO_MASSIMO_ATTREZZI) {
			if ( this.attrezzi[i] != null && this.attrezzi[i].getNome().equals(nomeAttrezzo)) {
				attrezzo = this.attrezzi[i];
				this.numeroAttrezzi -= 1;
				this.attrezzi[i] = null;
			}
			i++;
		}
		return attrezzo;
	}

	public String toString() {
		StringBuilder s = new StringBuilder();
		if (!this.isEmpty()) {
			s.append("Contenuto borsa ("+this.getPeso()+"kg/"+this.getPesoMax()+"kg): ");
			for (int i= 0; i<this.numeroAttrezzi; i++)
				s.append(attrezzi[i].toString()+" ");
		}
		else 
			s.append("Borsa vuota");
		return s.toString();
	}
}